rename Leanplum-Internal back to LeanplumInternal to avoid breaking other modules
