+++ Leanplum-SDK/Classes/LPInbox.m
@@ -25,10 +25,10 @@
 #import "LPInbox.h"
 #import "Constants.h"
 #import "Leanplum.h"
-#import "Leanplum-Internal.h"
+#import "LeanplumInternal.h"
 #import "LeanplumRequest.h"
 #import "LPVarCache.h"
-#import "Leanplum-Internal.h"
+#import "LeanplumInternal.h"
 #import "LPAES.h"
 #import "LPKeychainWrapper.h"
 #import "LPFileManager.h"
